requirements.txt
anyascii==0.3.1
asgiref==3.4.1
atomicwrites==1.4.1
attrs==22.2.0
autopep8==1.4.4
beautifulsoup4==4.11.2
bleach==3.3.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.6
defusedxml==0.6.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==2.2.14
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-environ==0.8.1
django-filter==22.1
django-localflavor==3.1
django-mailer==2.2
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.6.0
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.0.4
html5lib==1.1
idna==2.8
importlib-metadata==6.0.0
l18n==2021.3
lockfile==0.12.2
Markdown==2.6.6
more-itertools==9.1.0
oauthlib==3.0.1
openpyxl==3.1.1
optional-django==0.1.0
packaging==23.0
pep8==1.7.1
Pillow==6.2.2
pluggy==0.13.1
py==1.11.0
pycodestyle==2.5.0
pytest==5.4.1
pytest-django==3.9.0
python-decouple==3.1
python-dotenv==0.12.0
python-stdnum==1.18
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.14.0
soupsieve==2.4
sqlparse==0.2.4
static3==0.7.0
stripe==2.27.0
telepath==0.3
toml==0.10.2
typing_extensions==4.5.0
Unipath==1.0
urllib3==1.24.2
wagtail==4.2
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.4.1
wincertstore==0.2
zipp==3.15.0
